03-prefix-caching/requirements.txt
httpx>=0.28.0
pandas>=2.2.0
matplotlib>=3.9.0
pyyaml>=6.0
tqdm>=4.67.0
pytest>=8.0.0
pytest-asyncio>=0.24.0
# GPU-only:
# vllm>=0.16.0
# sglang[all]>=0.5.0
